# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Yone wi",
        "description": "Cet api permet de gerer des reseaux de transports en commun au senegal et le partage d'informations sur les les abonnements ,trajets et leurs tarifs  ",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/tarifs": {
            "get": {
                "tags": [
                    "Gestion des tarifs"
                ],
                "summary": "Lister les tarifs",
                "description": "",
                "operationId": "a8d068b207aa887701e2c48354d0964c",
                "parameters": [
                    {
                        "name": "User-Agent",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Gestion des tarifs"
                ],
                "summary": "Ajouter un tarif",
                "description": "",
                "operationId": "2a2fd913fc8d15e7635aba1fd9651a7c",
                "parameters": [
                    {
                        "name": "User-Agent",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "prix": {
                                        "type": "integer"
                                    },
                                    "type": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created successfully"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorize"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/mestarifs": {
            "get": {
                "tags": [
                    "Gestion des tarifs"
                ],
                "summary": "Lister mes tarifs",
                "description": "",
                "operationId": "fa576a93103cf7b7a78794573b0cbb8a",
                "parameters": [
                    {
                        "name": "User-Agent",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/tarifs/{tarif}": {
            "get": {
                "tags": [
                    "Gestion des tarifs"
                ],
                "summary": "Afficher un tarif",
                "description": "",
                "operationId": "ac01b38322914b3bc69d454244a716f4",
                "parameters": [
                    {
                        "name": "tarif",
                        "in": "path",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "User-Agent",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Gestion des tarifs"
                ],
                "summary": "Supprimer un tarif",
                "description": "",
                "operationId": "89704da317e4d962b60d2f238e13954f",
                "parameters": [
                    {
                        "name": "tarif",
                        "in": "path",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "User-Agent",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorize"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Gestion des tarifs"
                ],
                "summary": "Modifier un tarif",
                "description": "",
                "operationId": "ef56e1a2f989cfe57c3b18b2e07386b6",
                "parameters": [
                    {
                        "name": "tarif",
                        "in": "path",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "User-Agent",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "prix": {
                                        "type": "integer"
                                    },
                                    "type": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/tarifs/restaurer/{tarif}": {
            "patch": {
                "tags": [
                    "Gestion des tarifs"
                ],
                "summary": "restaurer un tarif",
                "description": "",
                "operationId": "46bd3a9b73e3342922d0edbd2c0ef2c0",
                "parameters": [
                    {
                        "name": "tarif",
                        "in": "path",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "User-Agent",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/tarifs/delete/{tarif}": {
            "patch": {
                "tags": [
                    "Gestion des tarifs"
                ],
                "summary": "supprimer un tarif de la corbeille",
                "description": "",
                "operationId": "c1ac171a0b31a4f63600aab2e7038a3d",
                "parameters": [
                    {
                        "name": "tarif",
                        "in": "path",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "User-Agent",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/tarifs/deleted/all": {
            "get": {
                "tags": [
                    "Gestion des tarifs"
                ],
                "summary": "Lister les tarifs qui sont dans la corbeille",
                "description": "",
                "operationId": "bd9453bcbe33c3c6f615de82c9f4e5f9",
                "parameters": [
                    {
                        "name": "User-Agent",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/api/tarifs/empty-trash": {
            "post": {
                "tags": [
                    "Gestion des tarifs"
                ],
                "summary": "vider les tarifs qui sont dans la corbeille",
                "description": "",
                "operationId": "26648c98d31b4d8c70ccf715f8d81d6e",
                "parameters": [
                    {
                        "name": "User-Agent",
                        "in": "header",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created successfully"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorize"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}